refactor(DES-3210):building a package structure
