+++ src/main/java/com/gbtec/initialization/GenerationRandomizer.java
@@ -1,4 +1,6 @@
-package com.gbtec.Initialization;
+package com.gbtec.initialization;
+
+import com.gbtec.model.Generation;
 
 import java.util.Random;
