+++ src/main/java/com/gbtec/model/Cell.java
@@ -1,4 +1,4 @@
-package com.gbtec.Initialization;
+package com.gbtec.model;
 
 public class Cell {
